closes #28
